cleanup a11y addon
