+++ addons/a11y/src/constants.ts
@@ -1,7 +1,6 @@
 export const ADDON_ID = 'storybook/a11y';
 export const PANEL_ID = `${ADDON_ID}/panel`;
 export const PARAM_KEY = `a11y`;
-export const IFRAME = 'iframe';
 export const HIGHLIGHT_STYLE_ID = 'a11yHighlight';
 const RESULT = `${ADDON_ID}/result`;
 const REQUEST = `${ADDON_ID}/request`;
